Proper renames: remove SlickEngagement_ prefix from ActionsFilters, Init, InstallIndicator, PageBootData, Plugin, PluginInit and update references
