+++ svn/trunk/ActionsFilters.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Slickstream;
 
 require_once 'SlickEngagement_LifeCycle.php';
-require_once 'SlickEngagement_Plugin.php';
+require_once 'Plugin.php';
 
 const GENESIS_AFTER_HEADER_POSTS = 'After header on posts (for Genesis themes)';
 const GENESIS_BEFORE_CONTENT_POSTS = 'Before content on posts (for Genesis themes)';
