+++ svn/trunk/Init.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 function SlickEngagement_init(): void
 {
 
-    require_once 'SlickEngagement_PluginInit.php';
-    require_once 'SlickEngagement_ActionsFilters.php';
+    require_once 'PluginInit.php';
+    require_once 'ActionsFilters.php';
     $slickActionsFilters = new Slickstream\ActionsFilters();
 
     // NOTE: this file gets run each time you *activate* the plugin.
