+++ svn/trunk/InstallIndicator.php
