+++ svn/trunk/PluginInit.php
@@ -7,6 +7,6 @@ namespace Slickstream;
 //Added for backwards compatibility; do not remove/rename
 function PluginInit(): void
 {
-    require_once 'SlickEngagement_Init.php';
+    require_once 'Init.php';
     \SlickEngagement_init();
 }
